refactor #51: IssueList에서 IssueTable로 함수명 수정
