+++ fe/src/components/IssueTable/Issue.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Profile from '@common/Profile';
 import Label from '@common/Label';
-import { LabelRow } from './IssueList';
+import { LabelRow } from './IssueTable';
 import { ElapseTime } from '@utils/getTimeElapsed';
 import { ReactComponent as AlertCircle } from '@assets/alertCircle.svg';
 import { ReactComponent as Archive } from '@assets/archive.svg';
